+++ test/libsolidity/syntaxTests/metaTypes/address.sol
@@ -1,0 +1,8 @@
+contract Test {
+    function basic() public pure {
+        address a = type(address).min;
+        a;
+        address b = type(address).max;
+        b;
+    }
+}